x/proof/keeper/msg_update_params_test.go
package keeper_test

import (
	"testing"

	"github.com/stretchr/testify/require"

	"github.com/pokt-network/poktroll/proto/types/proof"
)

func TestMsgUpdateParams(t *testing.T) {
	k, ms, ctx := setupMsgServer(t)
	params := proof.DefaultParams()
	require.NoError(t, k.SetParams(ctx, params))

	// default params
	tests := []struct {
		desc           string
		params         *proof.MsgUpdateParams
		shouldError    bool
		expectedErrMsg string
	}{
		{
			desc: "invalid: authority address invalid",
			params: &proof.MsgUpdateParams{
				Authority: "invalid",
				Params:    params,
			},
			shouldError:    true,
			expectedErrMsg: "invalid authority",
		},
		{
			desc: "invalid: send empty params",
			params: &proof.MsgUpdateParams{
				Authority: k.GetAuthority(),
				Params:    proof.Params{},
			},
			shouldError: true,
		},
		{
			desc: "valid: send minimal params",
			params: &proof.MsgUpdateParams{
				Authority: k.GetAuthority(),
				Params: proof.Params{
					ProofMissingPenalty: &proof.DefaultProofMissingPenalty,
				},
			},
			shouldError: false,
		},
		{
			desc: "valid: send default params",
			params: &proof.MsgUpdateParams{
				Authority: k.GetAuthority(),
				Params:    params,
			},
			shouldError: false,
		},
	}

	for _, test := range tests {
		t.Run(test.desc, func(t *testing.T) {
			_, err := ms.UpdateParams(ctx, test.params)

			if test.shouldError {
				require.Error(t, err)
				require.Contains(t, err.Error(), test.expectedErrMsg)
			} else {
				require.NoError(t, err)
			}
		})
	}
}
